### 2021-11-04 #3889 @whedon
[PRE REVIEW]: excluder: An R package that checks for exclusion criteria in online data
**Submitting author:** @JeffreyRStevens (<a href="http://orcid.org/0000-0003-2375-1360">Jeffrey R Stevens</a>)
**Repository:** <a href="https://github.com/ropensci/excluder/" target ="_blank">https://github.com/ropensci/excluder/</a>
**Version:** v0.3.1
