### 2021-11-05 #3889 @JeffreyRStevens
Thanks, @mikldk. The package has already been peer reviewed in rOpenSci by Joseph O'Brien (@jmobrien) and Julia Silge (@juliasilge) at https://github.com/ropensci/software-review/issues/455. Is there a fast-track review option since it has already been peer reviewed? If so, do you still need additional reviewers?
Thanks.
