### 2021-11-05 #3893 @danielskatz
Accepted by rOpenSci in https://github.com/ropensci/software-review/issues/455
