### 2021-11-05 #3893 @danielskatz
@JeffreyRStevens - is there already an archived version of this software in zenodo or a similar archival repository as part of the rOpenSci review?  If so, what's the DOI and version of the software that was archived?
