### 2021-11-05 #3893 @danielskatz
Yes - the paper looks good to me, so the next steps are:
- [ ] Make a tagged release of your software if needed, and list the version tag of the archived version here.
