### 2021-11-05 #3893 @JeffreyRStevens
* I have tagged the submitted version 0.3.1 on GitHub.
* I have archived the software on Zenodo at https://doi.org/10.5281/zenodo.5648202 with the DOI of 10.5281/zenodo.5648202.

Was that everything?
